Added original file storage and fetching

+++ main.py
@@ -1,3 +1,4 @@
+import glob
 import json
 import os
 import subprocess
@@ -7,22 +8,22 @@ from fastapi import FastAPI, Request, UploadFile, Form
 from fastapi.responses import JSONResponse, FileResponse
 from starlette.responses import HTMLResponse
 
-import db
+import Db
 from BaseModels.Films import FilmsBaseModel
 from Entities.Film import Film, FilmType, FilmFormat
 from Entities.FilmRoll import FilmRoll, DevelopmentStatus
 from Entities.Picture import Picture
-from config.config import Config
-from definitions import mime_file_extension
+from Config.config import Config
+from definitions import mime_file_extension, raw_file_types
 
 app = FastAPI(title="FilmStore",
               description="""An API to manage Film rolls, stocks and the images you shot on them.""",
               version="1.0.0",
               license_info={
-                "name": "GNU GPLv3",
-                "url": "https://gnu.org/copyright/",
+                  "name": "GNU GPLv3",
+                  "url": "https://gnu.org/copyright/",
               })
-config = Config(open(file="./config/config.json", mode="r"))
+config = Config(open(file="Config/config.json", mode="r"))
 
 
 @app.get("/")
@@ -52,10 +53,18 @@ async def api_root():
     })
 
 
+@app.get("/api/v1/config")
+async def get_api_config():
+    return JSONResponse(
+        status_code=200,
+        content=config.to_dict()
+    )
+
+
 @app.get("/api/v1/film_stocks")
 async def list_films():
     try:
-        films = db.film_stocks.fetch_all()
+        films = Db.film_stocks.fetch_all()
     except Exception as e:
         print(e)
         return JSONResponse(status_code=500, content={
@@ -74,7 +83,7 @@ async def list_films():
 @app.get("/api/v1/film_stocks/{film_id}")
 async def get_film(film_id: int):
     try:
-        film = db.film_stocks.fetch(film_id=film_id)
+        film = Db.film_stocks.fetch(film_id=film_id)
     except Exception as e:
         print(e)
         return JSONResponse(status_code=500, content={
@@ -91,7 +100,7 @@ async def get_film(film_id: int):
 @app.get("/api/v1/film_rolls")
 async def list_filmrolls(stock: int = 0):
     try:
-        film_rolls = db.film_rolls.fetch_all(stock_filter=stock)
+        film_rolls = Db.film_rolls.fetch_all(stock_filter=stock)
     except Exception as e:
         print(e)
         return JSONResponse(status_code=500, content={
@@ -108,7 +117,7 @@ async def list_filmrolls(stock: int = 0):
 @app.get("/api/v1/film_rolls/{filmrollid}")
 async def get_film_roll(roll_id: int):
     try:
-        film_roll = db.film_rolls.fetch(filmroll_id=roll_id)
+        film_roll = Db.film_rolls.fetch(filmroll_id=roll_id)
     except Exception as e:
         return JSONResponse(status_code=500, content={"success": False, "error": str(e)})
 
@@ -126,7 +135,7 @@ async def get_film_roll(roll_id: int):
 @app.get("/api/v1/pictures")
 async def list_pictures():
     try:
-        pictures = db.pictures.fetch_all()
+        pictures = Db.pictures.fetch_all()
     except Exception as e:
         return JSONResponse(status_code=500, content={"success": False, "error": str(e)})
 
@@ -144,7 +153,7 @@ async def list_pictures():
 @app.get("/api/v1/pictures/{picture_id}")
 async def get_picture(picture_id: int):
     try:
-        picture = db.pictures.fetch(picture_id=picture_id)
+        picture = Db.pictures.fetch(picture_id=picture_id)
     except Exception as e:
         return JSONResponse(status_code=500, content={"success": False, "error": str(e)})
     return JSONResponse(status_code=200, content={
@@ -167,17 +176,41 @@ async def get_picture(filename: str):
                         path=f'./pictures/{filename}')
 
 
+@app.get("/api/v1/pictures/hires/{filename}")
+async def get_picture(filename: str):
+    if config.original_folder is None:
+        return JSONResponse(status_code=400,
+                            content={
+                                "success": False,
+                                "error": "Original file storage is disabled"
+                            })
+
+    files = glob.glob(f'{config.original_folder}{filename}.*')
+
+    if len(files) == 0:
+        return JSONResponse(status_code=404,
+                            content={
+                                "success": False,
+                                "error": "File not found"
+                            })
+    file = files[0]
+
+    return FileResponse(status_code=200,
+                        media_type=list(mime_file_extension.keys())[list(mime_file_extension.values()).index(file.split('.')[-1])][0],
+                        path=file)
+
+
 @app.post("/api/v1/film_stocks")
 async def create_film(request: FilmsBaseModel):
     film = Film(name=request.name,
                 iso=request.iso,
                 development_info=request.development_info,
                 type=FilmType(request.type),
-                format=FilmFormat(request.format),)
+                format=FilmFormat(request.format), )
 
     # TODO)) ADD id and response id
     try:
-        db_id = db.film_stocks.create(film=film)
+        db_id = Db.film_stocks.create(film=film)
     except Exception as e:
         return JSONResponse(status_code=500, content={
             "message": str(e)
@@ -190,25 +223,49 @@ async def create_film(request: FilmsBaseModel):
 
 @app.put("/api/v1/pictures")
 async def upload_image_file(request: UploadFile):
+    use_processing = True if request.headers["Filmstore-Postprocess-Negative"] in ["True", "true"] else False
+    postprocessing_type = request.headers["Filmstore-Postprocess-Type"]
+
+    if use_processing and not config.allow_raw_post_processing:
+        return JSONResponse(status_code=400, content={
+            "success": False,
+            "error": "Postprocessing is not allowed"
+        })
+
     if request.content_type not in mime_file_extension.keys():
-        tempfile_ext: str = mime_file_extension[request.content_type]
+        original_file_ext: str = mime_file_extension[request.content_type]
 
-        filename = f"{uuid.uuid4().hex}.{tempfile_ext}"
+        if not config.allow_raw_upload and original_file_ext in raw_file_types.values():
+            return JSONResponse(status_code=400, content={
+                "success": False,
+                "error": "RAW File Upload is disabled"
+            })
 
+        filename = f"{uuid.uuid4().hex}.{original_file_ext}"
+
+        # Store temporary file for thumbnail
         with open(f"{config.temporary_files_folder}{filename}", "wb") as f:
             f.write(request.file.read())
 
-        subprocess.run(
-            ["./scripts/img.sh", f"{config.temporary_files_folder}{filename}", f"./pictures/{filename.split('.')[0]}.jpg"],
-        )
+        # Store original file if allowed
+        if config.original_folder is not None:
+            with open(f"{config.original_folder}{filename}", "wb") as f:
+                f.write(request.file.read())
+
+        if original_file_ext in raw_file_types.values():
+
+
+        subprocess.run(["./Scripts/img.sh",
+                        f"{config.temporary_files_folder}{filename}",
+                        f"./pictures/{filename.split('.')[0]}.jpg"])
 
         os.remove(f".temp/{filename}")
 
-        return JSONResponse(status_code=200, content={
+        return JSONResponse(status_code=201, content={
             "success": True,
-            "image_filename": filename,
-            "thumbnail_path": f"/api/v1/pictures/file/{filename}",
-            "original_file": f"",
+            "image_filename": filename.split('.')[0],
+            "thumbnail_path": f"/api/v1/pictures/file/{filename.split('.')[0]}.jpg",
+            "original_file": f"/api/v1/pictures/hires/{filename}" if config.allow_raw_upload else None,
             "awaiting_metadata": True
         })
     else:
@@ -220,7 +277,7 @@ async def upload_image_file(request: UploadFile):
 
 @app.post("/api/v1/pictures")
 async def upload_picture(request):
-    req_json = await request.json()
+    req_json = await request._json()
 
     # else:  # TODO)) ADD FILM FORMAT
 
@@ -233,7 +290,7 @@ async def upload_picture(request):
                       printed=req_json["printed"])
 
     try:
-        insert_id = db.pictures.create(picture=picture)
+        insert_id = Db.pictures.create(picture=picture)
     except Exception as e:
         return JSONResponse(status_code=500, content={
             "success": False,
@@ -247,7 +304,7 @@ async def upload_picture(request):
 
 
 @app.post("/api/v1/film_rolls")
-async def add_filmroll(request: Request):
+async def add_film_roll(request: Request):
     req_json = await request.json()
 
     pictures = [Picture(thumbnail="", db_id=id) for id in req_json["pictures"]]
@@ -255,12 +312,12 @@ async def add_filmroll(request: Request):
 
     filmroll = FilmRoll(camera=req_json["camera"],
                         archival_identifier=req_json["identifier"],
-                        status= DevelopmentStatus(req_json["status"]),
+                        status=DevelopmentStatus(req_json["status"]),
                         pictures=pictures,
                         film=film)
 
     try:
-        id = db.film_rolls.create(filmroll=filmroll)
+        id = Db.film_rolls.create(filmroll=filmroll)
     except Exception as e:
         return JSONResponse(status_code=500, content={
             "success": False,
@@ -281,7 +338,7 @@ async def update_film(request: Request, id: int):
 @app.delete("/api/v1/film_stocks/{id}")
 async def delete_film(stock_id: int):
     try:
-        resp_json = db.film_stocks.delete(film_stock_id=stock_id)
+        resp_json = Db.film_stocks.delete(film_stock_id=stock_id)
     except Exception as e:
         return JSONResponse(status_code=500, content={"success": False, "error": str(e)})
